+++ examples/Follow/followFromSearch.php
@@ -13,7 +13,7 @@ $truncatedDebug = false;
 //////////////////////
 
 //////////////////////
-$usernameToFollow = 'selenagomez';
+$queryUser = 'selenagomez';
 //////////////////////
 
 $ig = new \InstagramAPI\Instagram($debug, $truncatedDebug);
@@ -48,7 +48,7 @@ try {
     $timeToSearch = mt_rand(2000, 3500);
     sleep($timeToSearch / 1000);
 
-    $searchResponse = $ig->discover->search($usernameToFollow);
+    $searchResponse = $ig->discover->search($queryUser);
     $ig->event->sendNavigation('button', 'explore_popular', 'blended_search');
 
     $searchResults = $searchResponse->getList();
@@ -61,7 +61,7 @@ try {
     foreach ($searchResults as $searchResult) {
         if ($searchResult->getUser() !== null) {
             $resultList[] = $searchResult->getUser()->getPk();
-            if ($searchResult->getUser()->getUsername() === $usernameToFollow) {
+            if ($searchResult->getUser()->getUsername() === $queryUser) {
                 $found = true;
                 $userId = $searchResult->getUser()->getPk();
             }
@@ -77,8 +77,37 @@ try {
             $position++;
         }
     }
-    $ig->event->sendSearchResults($usernameToFollow, $resultList, $resultTypeList, $rankToken, $searchSession, 'blended_search');
-    $ig->event->sendSearchResultsPage($usernameToFollow, $userId, $resultList, $resultTypeList, $rankToken, $searchSession, $position, 'USER', 'blended_search');
+
+    $ig->event->sendNavigation('button', 'blended_search', 'blended_search');
+
+    $ig->event->sendSearchResults($queryUser, $resultList, $resultTypeList, $rankToken, $searchSession, 'blended_search');
+    $ig->event->sendSearchResultsPage($queryUser, $userId, $resultList, $resultTypeList, $rankToken, $searchSession, $position, 'USER', 'blended_search');
+
+    // When we clicked the user, we are navigating from 'blended_search' to 'profile'.
+    $ig->event->sendNavigation('search_result', 'blended_search', 'profile', null, null,
+        [
+            'rank_token'            => null,
+            'query_text'            => $queryUser,
+            'search_session_id'     => $searchSession,
+            'selected_type'         => 'user',
+            'position'              => 0,
+            'username'              => $queryUser,
+            'user_id'               => $userId,
+        ]
+    );
+    $ig->event->sendNavigation('button', 'profile', 'profile', null, null,
+        [
+            'rank_token'            => null,
+            'query_text'            => $queryUser,
+            'search_session_id'     => $searchSession,
+            'selected_type'         => 'user',
+            'position'              => 0,
+            'username'              => $queryUser,
+            'user_id'               => $userId,
+            'class_selector'        => 'ProfileMediaTabFragment',
+        ]
+    );
+
     $ig->discover->registerRecentSearchClick('user', $userId);
     $ig->people->getFriendship($userId);
     $suggestions = $ig->people->getInfoById($userId, 'search_users')->getUser()->getChainingSuggestions();
@@ -89,17 +118,7 @@ try {
             $ig->event->sendSimilarEntityImpression($userId, $suggestions[$i]->getPk());
         }
     }
-    $ig->event->sendNavigation('button', 'search_users', 'profile', null, null,
-        [
-            'rank_token'        => $rankToken,
-            'query_text'        => $usernameToFollow,
-            'search_session_id' => $searchSession,
-            'selected_type'     => 'user',
-            'position'          => 0,
-            'username'          => $usernameToFollow,
-            'user_id'           => $userId,
-        ]
-    );
+
     $traySession = \InstagramAPI\Signatures::generateUUID();
     $ig->highlight->getUserFeed($userId);
     $ig->story->getUserStoryFeed($userId);
@@ -134,10 +153,6 @@ try {
             [
                 'module'        => 'explore_popular',
                 'click_point'   => 'button',
-            ],
-            [
-                'module'        => 'explore_popular',
-                'click_point'   => 'explore_topic_load',
             ],
             [
                 'module'        => 'feed_timeline',